requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
werkzeug
sqlalchemy
python-dotenv
email-validator
pytz
flask-migrate
psycopg2-binary
flask-mongoengine
pymongo
mongoengine
dnspython
pywebview
pyinstaller
gunicorn
